Refactor encryption algorithms into own namespace
